Standardize and cleanup product data construction
