+++ src/models/snipcart.ts
@@ -1,0 +1,11 @@
+/**
+ * Data required for Snipcart validation.
+ */
+export interface ISnipcartProduct {
+  id: string;
+  name: string;
+  price: number;
+  url: string;
+  description: string;
+  image: string;
+}
